TotalPrice in Cart changed to decimal
TotalPrice in Cart changed to decimal

+++ FoodShop.Web/FoodShop.Data.Models/Cart.cs
@@ -21,6 +21,6 @@ namespace FoodShop.Data.Models
 
         public ICollection<CartItem> CartItems { get; set; }
 
-        public int Totalprice { get; set; }
+        public decimal Totalprice { get; set; }
     }
 }
